Make sure nanovg context always exists by drawing welcome panel with nanovg, added fade to recently opened viewport
