+++ Source/Canvas.h
@@ -60,7 +60,6 @@ class Canvas : public Component
     , public pd::MessageListener
     , public AsyncUpdater 
     , public NVGComponent
-    , public NVGContextListener
 {
 public:
     Canvas(PluginEditor* parent, pd::Patch::Ptr patch, Component* parentGraph = nullptr);
@@ -100,8 +99,6 @@ public:
     void synchronise();
     void performSynchronise();
     void handleAsyncUpdate() override;
-
-    void lookAndFeelChanged() override;
 
     void updateDrawables();
 
@@ -157,8 +154,6 @@ public:
 
     void orderConnections();
     
-    void nvgContextDeleted(NVGcontext* nvg) override;
-
     void showSuggestions(Object* object, TextEditor* textEditor);
     void hideSuggestions();
