+++ Source/Objects/GraphOnParent.h
@@ -25,7 +25,6 @@ public:
     GraphOnParent(pd::WeakReference obj, Object* object)
         : ObjectBase(obj, object)
         , subpatch(new pd::Patch(obj, cnv->pd, false))
-        , textRenderer(cnv->editor->nvgSurface)
     {
         resized();
