+++ Source/Tabbar/Tabbar.cpp
@@ -16,246 +16,6 @@
 #include "Utility/StackShadow.h"
 #include "Utility/Autosave.h"
 
-class WelcomePanel : public Component {
-
-    class WelcomePanelTile : public Component
-    {
-        String tileName;
-        String tileSubtitle;
-        float snapshotScale;
-        bool isHovered = false;
-        std::unique_ptr<Drawable> snapshot = nullptr;
-    public:
-        
-        std::function<void()> onClick = [](){};
-        
-        WelcomePanelTile(String name, String subtitle, String svgImage, Colour iconColour, float scale)
-        : tileName(name)
-        , tileSubtitle(subtitle)
-        , snapshotScale(scale)
-        {
-            snapshot = Drawable::createFromImageData(svgImage.toRawUTF8(), svgImage.getNumBytesAsUTF8());
-            if(snapshot) {
-                snapshot->replaceColour (Colours::black, iconColour);
-            }
-            
-            resized();
-        }
-        
-        void paint(Graphics& g) override
-        {
-            auto bounds = getLocalBounds().reduced(12);
-            
-            Path tilePath;
-            tilePath.addRoundedRectangle(bounds.getX(), bounds.getY(), bounds.getWidth(), bounds.getHeight(), Corners::largeCornerRadius);
-            
-            StackShadow::renderDropShadow(g, tilePath, Colour(0, 0, 0).withAlpha(0.1f), 7, {0, 2});
-            g.setColour(findColour(PlugDataColour::canvasBackgroundColourId));
-            g.fillPath(tilePath);
-            
-            if(snapshot)  {
-                snapshot->drawAt(g, 0, 0, 1.0f);
-            }
-            
-            Path textAreaPath;
-            textAreaPath.addRoundedRectangle(bounds.getX(), bounds.getHeight() - 32, bounds.getWidth(), 44, Corners::largeCornerRadius, Corners::largeCornerRadius, false, false, true, true);
-        
-            auto hoverColour = findColour(PlugDataColour::toolbarHoverColourId).interpolatedWith(findColour(PlugDataColour::toolbarBackgroundColourId), 0.5f);
-            g.setColour(isHovered ? hoverColour : findColour(PlugDataColour::toolbarBackgroundColourId));
-            g.fillPath(textAreaPath);
-            
-            g.setColour(findColour(PlugDataColour::toolbarOutlineColourId));
-            g.strokePath(tilePath, PathStrokeType(1.0f));
-            
-            Fonts::drawStyledText(g, tileName, bounds.getX() + 10, bounds.getHeight() - 30, bounds.getWidth() - 8, 24, findColour(PlugDataColour::panelTextColourId), Semibold, 14);
-            
-            g.setColour(findColour(PlugDataColour::panelTextColourId).withAlpha(0.75f));
-            g.setFont(Fonts::getCurrentFont().withHeight(13.5f));
-            g.drawText(tileSubtitle, bounds.getX() + 10, bounds.getHeight() - 10, bounds.getWidth() - 8, 16, Justification::centredLeft);
-        }
-        
-        void mouseEnter(const MouseEvent& e) override
-        {
-            isHovered = true;
-            repaint();
-        }
-        
-        void mouseExit(const MouseEvent& e) override
-        {
-            isHovered = false;
-            repaint();
-        }
-        
-        void mouseUp(const MouseEvent& e) override
-        {
-            onClick();
-        }
-        
-        void resized() override
-        {
-            if(snapshot) {
-                auto bounds = getLocalBounds().reduced(12).withTrimmedBottom(44);
-                snapshot->setTransformToFit(bounds.withSizeKeepingCentre(bounds.getWidth() * snapshotScale, bounds.getHeight() * snapshotScale).toFloat(), RectanglePlacement::centred);
-            }
-        }
-    };
-    
-public:
-    WelcomePanel()
-    {
-        recentlyOpenedViewport.setViewedComponent(&recentlyOpenedComponent);
-        recentlyOpenedViewport.setScrollBarsShown(true, false, false, false);
-        recentlyOpenedComponent.setVisible(true);
-        addAndMakeVisible(recentlyOpenedViewport);
-        
-        // Update async to make sure silhouettes are available
-        MessageManager::callAsync([_this = SafePointer(this)](){
-            _this->update();
-        });
-    }
-
-    void resized() override
-    {
-        auto bounds = getLocalBounds().reduced(24).withTrimmedTop(36);
-        auto rowBounds = bounds.removeFromTop(160);
-        
-        const int desiredTileWidth = 190;
-        const int tileSpacing = 4;
-
-        int totalWidth = bounds.getWidth();
-        // Calculate the number of columns that can fit in the total width
-        int numColumns = totalWidth / (desiredTileWidth + tileSpacing);
-        // Adjust the tile width to fit within the available width
-        int actualTileWidth = (totalWidth - (numColumns - 1) * tileSpacing) / numColumns;
-        
-        if(newPatchTile) newPatchTile->setBounds(rowBounds.removeFromLeft(actualTileWidth));
-        rowBounds.removeFromLeft(4);
-        if(openPatchTile) openPatchTile->setBounds(rowBounds.removeFromLeft(actualTileWidth));
-        
-        bounds.removeFromTop(16);
-        
-        recentlyOpenedViewport.setBounds(getLocalBounds().withTrimmedTop(260));
-
-        int numRows = (tiles.size() + numColumns - 1) / numColumns;
-        int totalHeight = numRows * 160;
-
-        auto scrollable = Rectangle<int>(24, 0, totalWidth + 24, totalHeight);
-        recentlyOpenedComponent.setBounds(scrollable);
-
-        // Start positioning the tiles
-        rowBounds = scrollable.removeFromTop(160);
-        for (auto* tile : tiles)
-        {
-            if (rowBounds.getWidth() < actualTileWidth) {
-                rowBounds = scrollable.removeFromTop(160);
-            }
-            tile->setBounds(rowBounds.removeFromLeft(actualTileWidth));
-            rowBounds.removeFromLeft(tileSpacing);
-        }
-    }
-    
-    void update()
-    {
-        newPatchTile = std::make_unique<WelcomePanelTile>("New Patch", "Create a new empty patch", newIcon, findColour(PlugDataColour::panelTextColourId), 0.33f);
-        openPatchTile = std::make_unique<WelcomePanelTile>("Open Patch", "Browse for a patch to open", openIcon, findColour(PlugDataColour::panelTextColourId), 0.33f);
-        
-        newPatchTile->onClick = [this]() { newTab(); };
-        openPatchTile->onClick = [this](){ openPatch(); };
-        
-        addAndMakeVisible(*newPatchTile);
-        addAndMakeVisible(*openPatchTile);
-        
-        tiles.clear();
-
-        auto settingsTree = SettingsFile::getInstance()->getValueTree();
-        auto recentlyOpenedTree = settingsTree.getChildWithName("RecentlyOpened");
-        
-        auto snapshotColour = LookAndFeel::getDefaultLookAndFeel().findColour(PlugDataColour::objectSelectedOutlineColourId).withAlpha(0.3f);
-        
-        if (recentlyOpenedTree.isValid()) {
-            for (int i = 0; i < recentlyOpenedTree.getNumChildren(); i++) {
-                auto path = File(recentlyOpenedTree.getChild(i).getProperty("Path").toString());
-                auto silhoutteSvg = Autosave::findPatchSilhouette(path);
-                
-                if(silhoutteSvg.isEmpty() && path.existsAsFile())
-                {
-                    silhoutteSvg = OfflineObjectRenderer::patchToSVGFast(path.loadFileAsString());
-                }
-                
-                auto openTime = Time(static_cast<int64>(recentlyOpenedTree.getChild(i).getProperty("Time")));
-                auto diff = Time::getCurrentTime() - openTime;
-                String date;
-                if(diff.inDays() == 0) date = "Today";
-                else if(diff.inDays() == 1) date = "Yesterday";
-                else date = openTime.toString(true, false);
-                String time = openTime.toString(false, true, false, true);
-                String timeDescription = date + ", " + time;
-                
-                auto* tile = tiles.add(new WelcomePanelTile(path.getFileName(), timeDescription, silhoutteSvg, snapshotColour, 1.0f));
-                tile->onClick = [this, path](){
-                    onPatchOpen(path);
-                };
-                recentlyOpenedComponent.addAndMakeVisible(tile);
-            }
-        }
-
-        resized();
-    }
-
-    void show()
-    {
-        update();
-        setVisible(true);
-    }
-
-    void hide()
-    {
-        setVisible(false);
-    }
-    
-    void paint(Graphics& g) override
-    {
-        g.fillAll(findColour(PlugDataColour::panelBackgroundColourId));
-
-        Fonts::drawStyledText(g, "Welcome to plugdata", 32, 12, getWidth(), 40, findColour(PlugDataColour::panelTextColourId), Bold, 30, Justification::centredLeft);
-        //Fonts::drawStyledText(g, "Open a file to begin patching", 32, 42, getWidth(), 40, findColour(PlugDataColour::panelTextColourId), Thin, 23, Justification::centredLeft);
-        
-        Fonts::drawStyledText(g, "Recently opened", 32, 228, getWidth(), 32, findColour(PlugDataColour::panelTextColourId), Bold, 24, Justification::centredLeft);
-    }
-    
-    void lookAndFeelChanged() override
-    {
-        update();
-    }
-    
-    std::function<void(File)> onPatchOpen = [](File){};
-    std::function<void()> newTab = [](){};
-    std::function<void()> openPatch = [](){};
-    
-    static inline const String newIcon = "<?xml version=\"1.0\" standalone=\"no\"?>\n"
-    "<!DOCTYPE svg PUBLIC \"-//W3C//DTD SVG 1.1//EN\" \"http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd\" >\n"
-    "<svg xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" version=\"1.1\" viewBox=\"-10 0 2058 2048\">\n"
-    "   <path fill=\"currentColor\"\n"
-    "d=\"M1024 170v512q0 72 50 122t120 50h512v852q0 72 -50 122t-120 50h-1024q-70 0 -120 -50.5t-50 -121.5v-1364q0 -72 50 -122t120 -50h512zM1151 213l512 512h-469q-16 0 -29.5 -12.5t-13.5 -30.5v-469z\" />\n"
-    "</svg>\n";
-    
-    static inline const String openIcon = "<?xml version=\"1.0\" standalone=\"no\"?>\n"
-    "<!DOCTYPE svg PUBLIC \"-//W3C//DTD SVG 1.1//EN\" \"http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd\" >\n"
-    "<svg xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" version=\"1.1\" viewBox=\"-10 0 2058 2048\">\n"
-    "   <path fill=\"currentColor\"\n"
-    "d=\"M1180 555h506q72 0 126 47t64 118v13l2 14v768q0 76 -52 131t-128 60h-12h-1324q-76 0 -131 -51.5t-59 -127.5l-2 -12v-620l530 2l17 -2q51 -4 92 -33l4 -3t6 -5l4 -2zM700 342q59 0 109 32l14 11l181 149l-263 219l-8 4q-10 8 -24 11h-9h-530v-236q0 -76 52 -131\n"
-    "t128 -59h12h338z\" />\n"
-    "</svg>\n";
-
-    std::unique_ptr<WelcomePanelTile> newPatchTile;
-    std::unique_ptr<WelcomePanelTile> openPatchTile;
-    
-    Component recentlyOpenedComponent;
-    BouncingViewport recentlyOpenedViewport;
-    
-    OwnedArray<WelcomePanelTile> tiles;
-};
-
 class ButtonBar::GhostTab : public Component {
 public:
     explicit GhostTab(PlugDataLook& lnfRef)
@@ -526,22 +286,7 @@ TabComponent::TabComponent(PluginEditor* parent)
     newButton.onClick = [this]() {
         newTab();
     };
-
-    welcomePanel = std::make_unique<WelcomePanel>();
-    addAndMakeVisible(welcomePanel.get());
-
-    welcomePanel->newTab = [this]() {
-        newTab();
-    };
-
-    welcomePanel->openPatch = [this]() {
-        openProject();
-    };
-
-    welcomePanel->onPatchOpen = [this](File patchFile) {
-        openProjectFile(patchFile);
-    };
-
+    
     setVisible(false);
     setTabBarDepth(0);
     tabs->addMouseListener(this, true);
@@ -630,29 +375,24 @@ void TabComponent::moveTab(int currentIndex, int newIndex)
     tabs->moveTab(currentIndex, newIndex, true);
 }
 
-void TabComponent::openProject()
-{
-    editor->openProject();
-}
-
 void TabComponent::onTabChange(int tabIndex)
 {
     editor->updateCommandStatus();
-
+    
     // Show welcome panel if there are no tabs open
     if (tabs->getNumTabs() == 0) {
         setTabBarDepth(0);
         tabs->setVisible(false);
-        welcomePanel->show();
     } else {
         tabs->setVisible(true);
-        welcomePanel->hide();
         setTabBarDepth(30);
         // we need to update the dropzones, because no resize will be automatically triggered when there is a tab added from welcome screen
         if (auto* parentHolder = dynamic_cast<ResizableTabbedComponent*>(getParentComponent()))
             parentHolder->updateDropZones();
     }
 
+    editor->resized();
+    
     auto* cnv = getCurrentCanvas();
     if (!cnv || tabIndex == -1 || editor->pd->isPerformingGlobalSync)
         return;
@@ -665,6 +405,8 @@ void TabComponent::onTabChange(int tabIndex)
         if (tabBar->getCurrentCanvas())
             tabBar->getCurrentCanvas()->tabChanged();
     }
+    
+
 }
 
 void TabComponent::changeCallback(int newCurrentTabIndex, String const& newTabName)
@@ -691,15 +433,6 @@ void TabComponent::changeCallback(int newCurrentTabIndex, String const& newTabNa
     currentTabChanged(newCurrentTabIndex, newTabName);
 }
 
-void TabComponent::openProjectFile(File& patchFile)
-{
-    editor->autosave->checkForMoreRecentAutosave(patchFile, [this, patchFile]() {
-        editor->pd->loadPatch(URL(patchFile), editor);
-        SettingsFile::getInstance()->addToRecentlyOpened(patchFile);
-        editor->pd->titleChanged();
-    });
-}
-
 void TabComponent::setTabBarDepth(int newDepth)
 {
     if (tabDepth != newDepth) {
@@ -721,8 +454,6 @@ void TabComponent::handleAsyncUpdate()
 void TabComponent::resized()
 {
     auto content = getLocalBounds();
-
-    welcomePanel->setBounds(content);
     newButton.setBounds(3, 0, tabDepth, tabDepth); // slighly offset to make it centred next to the tabs
 
     auto tabBounds = content.removeFromTop(tabDepth).withTrimmedLeft(tabDepth);
